News_summarizer_project/requirements.txt
python-dotenv
openai
datetime
logging
requests
